Crate rpc

RPC (Remote Procedure Call)

A simple RPC implementation.

Example: Ping

Structs PingRequest and PingResponse are defined as examples to demonstrate the usage of the RPC library. They implement the RpcRequest and RpcResponse traits respectively.

This example demonstrates the basic usage of the RPC library by sending a ping request to a set of nodes and waiting for a response.

Start some docker containers with the internal docker network attached. And run the following command to start the example:

cargo run --example ping 172.19.0.2:16

Note that the IP address and port number are the address of the current node.

Or simply run the PowerShell script which handles everything for you:

.\examples\ping.ps1

A typical output of a node would be:

2024-03-16_15:24:28.099977Z 172.19.0.3:16 TRACE Logger initialized with default level: Trace
2024-03-16_15:24:28.100066Z 172.19.0.3:16 INFO Machine started with socket: 172.19.0.3:16
2024-03-16_15:24:28.100125Z 172.19.0.3:16 INFO Sleeping for 2 seconds to wait for peers to start
2024-03-16_15:24:28.100347Z 172.19.0.3:16 TRACE Listening on 172.19.0.3:16
2024-03-16_15:24:29.706205Z 172.19.0.3:16 TRACE Accepted connection from 172.19.0.2:35496
2024-03-16_15:24:29.706807Z 172.19.0.3:16 DEBUG Received request [Ping] from 172.19.0.2:35496
2024-03-16_15:24:29.707107Z 172.19.0.3:16 INFO Sent response [Pong] to 172.19.0.2:35496
2024-03-16_15:24:30.107307Z 172.19.0.3:16 TRACE Connected to 172.19.0.2:16
2024-03-16_15:24:30.107519Z 172.19.0.3:16 DEBUG Sent request [Ping] to 172.19.0.2:16
2024-03-16_15:24:30.107591Z 172.19.0.3:16 TRACE Connected to 172.19.0.4:16
2024-03-16_15:24:30.107689Z 172.19.0.3:16 DEBUG Sent request [Ping] to 172.19.0.4:16
2024-03-16_15:24:30.107765Z 172.19.0.3:16 INFO Received response [Pong] from 172.19.0.2:16
2024-03-16_15:24:30.107845Z 172.19.0.3:16 INFO Received response [Pong] from 172.19.0.4:16
2024-03-16_15:24:30.518019Z 172.19.0.3:16 TRACE Accepted connection from 172.19.0.4:39230
2024-03-16_15:24:30.518094Z 172.19.0.3:16 DEBUG Received request [Ping] from 172.19.0.4:39230
2024-03-16_15:24:30.518186Z 172.19.0.3:16 INFO Sent response [Pong] to 172.19.0.4:39230
2024-03-16_15:25:00.094457Z 172.19.0.3:16 INFO Task closed due to timeout: Elapsed(())

See the example ping at ./examples/ping for a complete example.

Structs

PingRequest

PingResponse

Service

RPC service.

Traits

RpcRequest

Trait for RPC request.

RpcResponse

Trait for RPC response.

Type Aliases

Byte

Bytes

PingService

Result