Struct rpc::v1::types::BlockTemplate

pub struct BlockTemplate {
    pub version: u32,
    pub rules: Option<Vec<String>>,
    pub vbavailable: Option<HashMap<String, u32>>,
    pub vbrequired: Option<u32>,
    pub previousblockhash: H256,
    pub transactions: Vec<BlockTemplateTransaction>,
    pub coinbaseaux: Option<HashMap<String, String>>,
    pub coinbasevalue: Option<u64>,
    pub coinbasetxn: Option<BlockTemplateTransaction>,
    pub target: H256,
    pub mintime: Option<i64>,
    pub mutable: Option<Vec<String>>,
    pub noncerange: Option<String>,
    pub sigoplimit: Option<u32>,
    pub sizelimit: Option<u32>,
    pub weightlimit: Option<u32>,
    pub curtime: u32,
    pub bits: u32,
    pub height: u32,
}

Block template as described in: https://github.com/bitcoin/bips/blob/master/bip-0022.mediawiki https://github.com/bitcoin/bips/blob/master/bip-0023.mediawiki https://github.com/bitcoin/bips/blob/master/bip-0009.mediawiki#getblocktemplate_changes https://github.com/bitcoin/bips/blob/master/bip-0145.mediawiki

Fields
version: u32

The preferred block version

rules: Option<Vec<String>>

Specific block rules that are to be enforced

vbavailable: Option<HashMap<String, u32>>

Set of pending, supported versionbit (BIP 9) softfork deployments Keys: named softfork rules Values: identifies the bit number as indicating acceptance and readiness for given key

vbrequired: Option<u32>

Bit mask of versionbits the server requires set in submissions

previousblockhash: H256

The hash of previous (best known) block

transactions: Vec<BlockTemplateTransaction>

Contents of non-coinbase transactions that should be included in the next block

coinbaseaux: Option<HashMap<String, String>>

Data that should be included in the coinbase's scriptSig content Keys: ignored Values: value to be included in scriptSig

coinbasevalue: Option<u64>

Maximum allowable input to coinbase transaction, including the generation award and transaction fees (in Satoshis)

coinbasetxn: Option<BlockTemplateTransaction>

information for coinbase transaction

target: H256

The hash target

mintime: Option<i64>

The minimum timestamp appropriate for next block time in seconds since epoch (Jan 1 1970 GMT)

mutable: Option<Vec<String>>

List of ways the block template may be changed, e.g. 'time', 'transactions', 'prevblock'

noncerange: Option<String>

A range of valid nonces (constant 00000000ffffffff)

sigoplimit: Option<u32>

Limit of sigops in blocks

sizelimit: Option<u32>

Limit of block size

weightlimit: Option<u32>

Limit of block weight

curtime: u32

Current timestamp in seconds since epoch (Jan 1 1970 GMT)

bits: u32

Compressed target of next block

height: u32

The height of the next block