Struct rpc::v1::types::BlockTemplateTransaction

pub struct BlockTemplateTransaction {
    pub data: RawTransaction,
    pub txid: Option<H256>,
    pub hash: Option<H256>,
    pub depends: Option<Vec<u64>>,
    pub fee: Option<i64>,
    pub sigops: Option<i64>,
    pub weight: Option<i64>,
    pub required: bool,
}

Transaction data as included in BlockTemplate

Fields
data: RawTransaction

Transaction data encoded in hexadecimal

txid: Option<H256>

Transaction id encoded in little-endian hexadecimal

hash: Option<H256>

Hash encoded in little-endian hexadecimal (including witness data)

depends: Option<Vec<u64>>

Transactions before this one (by 1-based index in 'transactions' list) that must be present in the final block if this one is

fee: Option<i64>

Difference in value between transaction inputs and outputs (in Satoshis). For coinbase transactions, this is a negative Number of the total collected block fees (ie, not including the block subsidy). If key is not present, fee is unknown and clients MUST NOT assume there isn't one

sigops: Option<i64>

Total SigOps cost, as counted for purposes of block limits. If key is not present, sigop cost is unknown and clients MUST NOT assume it is zero.

weight: Option<i64>

Total transaction weight, as counted for purposes of block limits.

required: bool

If provided and true, this transaction must be in the final block