Struct rpc::v1::types::VerboseBlock

pub struct VerboseBlock {
    pub hash: H256,
    pub confirmations: i64,
    pub size: u32,
    pub strippedsize: u32,
    pub weight: u32,
    pub height: Option<u32>,
    pub version: u32,
    pub version_hex: String,
    pub merkleroot: H256,
    pub tx: Vec<H256>,
    pub time: u32,
    pub mediantime: Option<u32>,
    pub nonce: u32,
    pub bits: u32,
    pub difficulty: f64,
    pub chainwork: U256,
    pub previousblockhash: Option<H256>,
    pub nextblockhash: Option<H256>,
}

Verbose block information

Fields
hash: H256

Block hash

confirmations: i64

Number of confirmations. -1 if block is on the side chain

size: u32

Block size

strippedsize: u32

Block size, excluding witness data

weight: u32

Block weight

height: Option<u32>

Block height TODO: bitcoind always returns value, but we hold this value for main chain blocks only

version: u32

Block version

version_hex: String

Block version as hex

merkleroot: H256

Merkle root of this block

tx: Vec<H256>

Transactions ids

time: u32

Block time in seconds since epoch (Jan 1 1970 GMT)

mediantime: Option<u32>

Median block time in seconds since epoch (Jan 1 1970 GMT) TODO: bitcoind always returns value, but we can calculate this only if height(block) > 2

nonce: u32

Block nonce

bits: u32

Block nbits

difficulty: f64

Block difficulty

chainwork: U256

Expected number of hashes required to produce the chain up to this block (in hex)

previousblockhash: Option<H256>

Hash of previous block

nextblockhash: Option<H256>

Hash of next block