Struct rpc::MetaIoHandler

pub struct MetaIoHandler<T, S = Noop> where
    S: Middleware<T>,
    T: Metadata,  { /* fields omitted */ }

Request handler

By default compatible only with jsonrpc v2

Methods

impl<T> MetaIoHandler<T, Noop> where
T: Metadata,

pub fn with_compatibility(
compatibility: Compatibility
) -> MetaIoHandler<T, Noop>

Creates new MetaIoHandler compatible with specified protocol version.

impl<T, S> MetaIoHandler<T, S> where
S: Middleware<T>,
T: Metadata,

pub fn new(compatibility: Compatibility, middleware: S) -> MetaIoHandler<T, S>

Creates new MetaIoHandler

pub fn with_middleware(middleware: S) -> MetaIoHandler<T, S>

Creates new MetaIoHandler with specified middleware.

pub fn add_alias(&mut self, alias: &str, other: &str)

Adds an alias to a method.

pub fn add_method<F>(&mut self, name: &str, method: F) where
F: RpcMethodSimple,

Adds new supported asynchronous method

pub fn add_notification<F>(&mut self, name: &str, notification: F) where
F: RpcNotificationSimple,

Adds new supported notification

pub fn add_method_with_meta<F>(&mut self, name: &str, method: F) where
F: RpcMethod<T>,

Adds new supported asynchronous method with metadata support.

pub fn add_notification_with_meta<F>(&mut self, name: &str, notification: F) where
F: RpcNotification<T>,

Adds new supported notification with metadata support.

pub fn extend_with<F>(&mut self, methods: F) where
F: Into<HashMap<String, RemoteProcedure<T>, RandomState>>,

Extend this MetaIoHandler with methods defined elsewhere.

pub fn handle_request_sync(&self, request: &str, meta: T) -> Option<String>

Handle given request synchronously - will block until response is available. If you have any asynchronous methods in your RPC it is much wiser to use handle_request instead and deal with asynchronous requests in a non-blocking fashion.

pub fn handle_request(
&self,
request: &str,
meta: T
) -> Map<Either<FutureResult<Option<Response>, ()>, <S as Middleware<T>>::Future>, fn(Option<Response>) -> Option<String>>

Handle given request asynchronously.

pub fn handle_rpc_request(
&self,
request: Request,
meta: T
) -> <S as Middleware<T>>::Future

Handle deserialized RPC request.

pub fn handle_call(
&self,
call: Call,
meta: T
) -> Either<Box<Future<Item = Option<Output>, Error = ()> + 'static + Send>, FutureResult<Option<Output>, ()>>

Handle single call asynchronously.